Release/5.53.2

See merge request CG_Mobile/CG_Module_Android/Backend2/backend2!361

+++ backend2/src/main/java/com/cmoney/backend2/forumocean/service/api/schemas/v2/GroupBoardArticle.kt
@@ -18,9 +18,9 @@ data class GroupBoardArticle(
     @SerializedName("content")
     var articleContent: GroupBoardArticleContent?,
     @SerializedName("createTime")
-    var createTime: Int?,
+    var createTime: Long?,
     @SerializedName("modifyTime")
-    var modifyTime: Int?,
+    var modifyTime: Long?,
     @SerializedName("commentCount")
     var commentCount: Int?,
     @SerializedName("donation")
